GYM-211: Move routine and workout pages to separate folders
