+++ e2e/pages/routine/routine-details.page.ts
@@ -1,5 +1,5 @@
 import { Locator, Page } from '@playwright/test';
-import { BasePage } from './base.page';
+import { BasePage } from '../base.page';
 
 export class RoutineDetailsPage extends BasePage {
   url: string;
