+++ e2e/pages/routine/routine-form-base.page.ts
@@ -1,11 +1,14 @@
 import { Locator, Page } from '@playwright/test';
-import { routineData } from '../test-data/routine.data';
+import { routineData } from '../../test-data/routine.data';
+import { BasePage } from '../base.page';
 
-export class RoutineUpdatePage {
+export class RoutineFormBasePage extends BasePage {
+  url = '';
   routineName: Locator;
   saveRoutineButton: Locator;
 
-  constructor(private page: Page) {
+  constructor(protected page: Page) {
+    super(page);
     this.routineName = this.page.getByRole('textbox', { name: 'Routine name' });
     this.saveRoutineButton = this.page.getByTestId('save-routine-button');
   }
