+++ e2e/pages/routine/routine-update.page.ts
@@ -1,0 +1,14 @@
+import { Page } from '@playwright/test';
+import { RoutineFormBasePage } from './routine-form-base.page';
+
+export class RoutineUpdatePage extends RoutineFormBasePage {
+  url: string;
+
+  constructor(
+    protected page: Page,
+    private routineId: string,
+  ) {
+    super(page);
+    this.url = `/routines/routine-update/${routineId}`;
+  }
+}
