+++ e2e/pages/workout/workout-form.page.ts
@@ -1,9 +1,9 @@
 import { expect, Locator, Page } from '@playwright/test';
 import { WorkoutBasePage } from './workout-base.page';
-import { RoutineExercise } from '../test-data/exercises.data';
+import { RoutineExercise } from '../../test-data/exercises.data';
 
 export class WorkoutFormPage extends WorkoutBasePage {
-  url= '/my-training/new';
+  url = '/my-training/new';
   title: Locator;
 
   constructor(protected page: Page) {
